boto3_scripts/requirements.txt
appnope==0.1.2
astroid==2.11.2
asttokens==2.0.5
async-generator==1.10
asyncio==3.4.3
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
boto3==1.21.25
botocore==1.24.25
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.1.0
chromedriver-binary==104.0.5112.20.0
convertdate==2.4.0
cryptography==37.0.4
debugpy==1.5.1
decorator==5.1.1
dill==0.3.4
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
h11==0.13.0
hijri-converter==2.2.3
holidays==0.13
idna==3.3
ipykernel==6.9.2
ipython==8.1.1
isort==5.10.1
jedi==0.18.1
jmespath==1.0.0
jsonpickle==2.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.7.1
matplotlib-inline==0.1.3
mccabe==0.7.0
nest-asyncio==1.5.4
numpy==1.22.3
openpyxl==3.0.10
outcome==1.2.0
pandas==1.4.1
parso==0.8.3
pdfminer.six==20220524
pdfplumber==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pybrowsers==0.5.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
pylint==2.13.4
PyMeeus==0.5.11
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.28.1
s3transfer==0.5.2
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
Wand==0.6.7
wcwidth==0.2.5
webdriver-manager==3.8.0
wrapt==1.14.0
wsproto==1.1.0
xlrd==2.0.1
